+++ snowddl/__init__.py
@@ -2,6 +2,7 @@ from .app import *
 from .blueprint import *
 from .config import SnowDDLConfig
 from .engine import SnowDDLEngine
+from .formatter import SnowDDLFormatter
 from .settings import SnowDDLSettings
 from .error import SnowDDLExecuteError, SnowDDLUnsupportedError
 from .parser import *